requirements.txt
python-dotenv
pandas
pinecone-client==2.2.4 
pinecone-text==0.8.0 
openai==1.9.0
llama-index==0.9.34
tqdm
